Add posix sh env script (#9)
This amounted taking the bash env script and removing its use of the
`local` keyword and changing is extension to ".sh". The posix env script
also doesn't source the bash completion script as it uses some
bash-specific features.

This allows the dune binary distro to be used on machines with minimal
shells such as the ash shell in the default environment on alpine linux.

Also adds a test that runs shellcheck on env.sh to make sure it stays
posix-compatible.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>