Add shell script that builds dune and makes tarball (#12)
This makes it easier to test the binary distro offline since the tarball
has the same logic as the release github action (which now just runs the
script). This also makes it more reliable to test the build and
packaging process since the script can be run from a test setting, so
this change also introduces a github action that runs such a test.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>