Fix paths in env scripts in nested shells (#13)
When starting a shell from an existing shell where the shell config has
already been sourced, or when sourcing a shell config for a second time
within the same shell session, any opam setup in the shell config will
re-insert the bin directory from the current opam switch at the
beginning of PATH.

Dune's setup logic doesn't re-insert directories already present in PATH
as dune may share its bin directory with other programs reliant on the
current PATH order. Thus opam's re-initialization will result in any
dune executable in the current opam switch taking precedence over the
binary distro.

To prevent this, the __dune_env function detects when it's called
subsequent times in a shell session, and if it succeeded the first time
in correctly setting up the environment for the dune binary distribution
will remove any opam directories from PATH which precede the first
instance of location of the dune binary distro in PATH. Opam bin
directories following the location of the binary distro in PATH, as well
as all other directories in PATH are left unchanged.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>