Various minor fixes and documentation (#17)
* Move completion url to flake.nix

Maintainers are expected to only modify flake.nix. While the url url of
the completion script isn't expected to change (unless Tarides adopts
the project), the goal of this change is to make it clear to maintainers
where the completion script comes from.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>

* Completion script is optional

Passing `completion = null;` to dune.nix causes the dune package to be
built without a completion script.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>

* Fix typos in comments

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>

* Add documentation to readme file

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>

---------

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>